==> cancer_detection_final/3B_inc_exc_ID.py <==
#!/usr/bin/env python
# coding: utf-8
#NOTE: use paimg9 env

import sys
import os
import torch
import pandas as pd
import warnings
sys.path.insert(0, '../Utils/')
from Preprocessing import combine_sampleinfo_and_label_all, get_cancer_detected_ids, concat_hr1_and_hr2_label
from Preprocessing import prepross_neptune_label_data, extract_before_second_underscore, preproposs_tcga_label, build_info_path, filter_cancer_ids
from misc_utils import create_dir_if_not_exists
warnings.filterwarnings("ignore")
import argparse


# source ~/.bashrc
# conda activate paimg9
#python3 -u 3B_inc_exc_ID.py --TUMOR_FRAC_THRES 0.8 --cohort_name OPX
############################################################################################################
#USER INPUT 
############################################################################################################
parser = argparse.ArgumentParser("Tile feature extraction")
parser.add_argument('--TUMOR_FRAC_THRES', default= 0.9, type=float, help='tile tumor fraction threshold')
parser.add_argument('--cohort_name', default='TCGA_PRAD', type=str, help='data set name: TAN_TMA_Cores or OPX or TCGA_PRAD or Neptune')
parser.add_argument('--out_folder', default= '3B_labels_final_sample', type=str, help='out folder name')

args = parser.parse_args()


############################################################################################################
#DIR
############################################################################################################
proj_dir = '/fh/fast/etzioni_r/Lucas/mh_proj/mutation_pred/'
wsi_location = proj_dir +  'data/' + args.cohort_name + "/"
label_path = os.path.join(proj_dir,'intermediate_data','3_labels', args.cohort_name)
info_path_st1_ol100 = build_info_path(proj_dir, args.cohort_name, 100, stnorm=True)
info_path_st1_ol0   = build_info_path(proj_dir, args.cohort_name, 0,   stnorm=True)
info_path_st0_ol100 = build_info_path(proj_dir, args.cohort_name, 100, stnorm=False)
info_path_st0_ol0   = build_info_path(proj_dir, args.cohort_name, 0,   stnorm=False)

out_location = os.path.join(proj_dir,'intermediate_data',
                            args.out_folder, 
                            args.cohort_name, 
                            'TFT' + str(args.TUMOR_FRAC_THRES))
create_dir_if_not_exists(out_location)

##################
#Select GPU
##################
device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')

    
    
if args.cohort_name.replace("z_nostnorm_", "") == "OPX":
    
    
    ################################################
    #Get OPX IDs 
    ################################################
    id_col = "SAMPLE_ID"
    
    #All Aval IDs
    opx_ids = [x.replace('.tif','') for x in os.listdir(wsi_location)] #360
    selected_ids = opx_ids
    
    #Include IDs are high quality
    label_df = pd.read_csv(os.path.join(label_path, "label_df_allslides.csv")) #274 Samples, 272 patient, #New data (there are some ids in old data exclude due to bad quality)
    ids_high_quality = list(label_df[id_col].unique()) 
    selected_ids = [x for x in selected_ids if x in ids_high_quality] #274
    

    #Exclude IDs that are in finetune train to exclude 
    fine_tune_ids_df = pd.read_csv(proj_dir + 'intermediate_data/0_cd_finetune/cancer_detection_training/all_tumor_fraction_info.csv')
    ft_train_ids = list(fine_tune_ids_df.loc[fine_tune_ids_df['Train_OR_Test'] == 'Train','sample_id']) #24, 7 from OPX, 17 from ccola
    selected_ids = [x for x in selected_ids if x not in ft_train_ids] #268
    
    # ############################################################################################################
    # #Combine site and label info and tile info
    # ############################################################################################################       
    # all_tile_info_df = combine_sampleinfo_and_label_all(info_path, label_df, selected_ids, id_col = 'SAMPLE_ID')  
    # print(all_tile_info_df.shape) #1743458 tiles overlap0, 4843073 tiles overlap100
    # print("Total OPX SP IDs in tile path: ", len(set(all_tile_info_df['SAMPLE_ID']))) #268,204
    # print("Total OPX PT IDs in tile path: ", len(set(all_tile_info_df['PATIENT_ID']))) #266,201

    # #Print stats
    # tile_counts = all_tile_info_df['SAMPLE_ID'].value_counts()
    # print("Max # tile/per sp:", tile_counts.max()) #34689, 96406
    # print("Min # tile/per sp:", tile_counts.min()) #98, 285
    # print("Median # tile/per sp:", tile_counts.median()) #1809.5,5132
    
elif args.cohort_name.replace("z_nostnorm_", "") == "TCGA_PRAD":    
    ################################################
    #Get TCGA IDs 
    ################################################
    id_col = "FOLDER_ID"
    
    #All Aval IDs
    tcga_ids = [x.replace('.svs','') for x in os.listdir(wsi_location) if x != '.DS_Store'] #449
    selected_ids = tcga_ids
    
    #Exclude tissue issue
    issue_ids = ['cca3af0c-3e0e-4cfb-bb07-459c979a0bd5']
    selected_ids = [x for x in selected_ids if x not in issue_ids]
    
    #Exclude IHC 
    issue_ids =['61887edb-c320-42bb-a0b5-fe84c17dbd21']
    selected_ids = [x for x in selected_ids if x not in issue_ids]
    
    #Include IDs are high quality
    label_df = pd.read_csv(os.path.join(label_path, "label_df_allslides.csv")) #274 Samples, 272 patient, #New data (there are some ids in old data exclude due to bad quality)
    ids_high_quality = list(label_df[id_col].unique()) 
    selected_ids = [x for x in selected_ids if x in ids_high_quality] #447
    
    
    

elif args.cohort_name.replace("z_nostnorm_", "") == "Neptune":
    ################################################
    #Get Neptune IDs 
    ################################################
    id_col = "SAMPLE_ID"
    
    #All Aval IDs
    nep_ids = [x.replace('.tif','') for x in os.listdir(wsi_location) if x != '.DS_Store'] #350
    selected_ids = nep_ids
    
    #Include IDs are high quality
    label_df = pd.read_csv(os.path.join(label_path, "label_df_allslides.csv")) #274 Samples, 272 patient, #New data (there are some ids in old data exclude due to bad quality)    
    ids_high_quality = list(label_df[id_col].unique()) 
    selected_ids = [x for x in selected_ids if x in ids_high_quality] #209
    
    



#Include Cancer detected IDs intersection betweeen all nst, st, ol100 and ol0
results = filter_cancer_ids(
    selected_ids,
    [
        info_path_st1_ol100,
        info_path_st1_ol0,
        info_path_st0_ol100,
        info_path_st0_ol0,
    ],
    threshold=  args.TUMOR_FRAC_THRES,
    id_col=id_col,
    cohort_name = args.cohort_name
)

common_ids = results["common_ids"]

#Exclude the non-cancer samples from all methods
selected_ids = list(common_ids)
selected_ids.sort()


#Updated label
label_df = label_df.loc[label_df[id_col].isin(selected_ids)]
print(f"Final {args.cohort_name} SAMPLE IDs (n = ", len(set(label_df['SAMPLE_ID'])), ")")
print(f"Final {args.cohort_name} PTS IDs (n = ", len(set(label_df['PATIENT_ID'])), ")")
label_df.to_csv(os.path.join(out_location, "final_sample_label_df.csv"), index = False)


# elif args.cohort_name.replace("z_nostnorm_", "") == 'TAN_TMA_Cores':
#     #All Aval IDs
#     tan_ids =  [x.replace('.tif','') for x in os.listdir(wsi_location)] #677
    
#     #Load TAN_TMA mutation label data
#     label_df1 = pd.read_excel(os.path.join(label_path, "TAN97_core_mappings.xlsx")) #These Ids not in label_df2: ['18-018', '18-087', '18-064', '18-077', '08-016', '06-131']
#     label_df1.rename(columns = {'AR': 'AR_inMappingFile'}, inplace = True)
#     label_df1.loc[pd.isna(label_df1['AR pos']),'AR pos'] = 0
#     label_df1.loc[pd.isna(label_df1['NE pos']),'NE pos'] = 0
#     label_df2 = pd.read_excel(os.path.join(label_path, "TAN_coded mutation_for Roman.xlsx"))
#     label_df2.rename(columns = {'AR coded': 'AR',
#                                'CHD1 coded': 'CHD1',
#                                'PTEN coded': 'PTEN',
#                                'RB1 coded': 'RB1',
#                                'TP53 coded': 'TP53', 
#                                'BRCA2 coded':'BRCA2'}, inplace = True)
#     #Combine to get core ids
#     #Only keep the ids in TAN_coded mutation_for Roman.xlsx, because no mutation labels are aviabale , cannot say it is negative
#     label_df = label_df1.merge(label_df2, left_on = ['ptid'], right_on = ['Sample'], how = 'right')
#     label_df.reset_index(drop=True, inplace=True)
        
#     # #There 40 sample IDs does not have matched AR status
#     # checkAR = label_df.loc[label_df['AR pos'] != label_df['AR'],]
#     # print(len(set(checkAR['Sample'])))
#     # checkAR.to_csv(out_location + "AR_notmatch.csv", index = False)
    
    
#     #Recode SITE info
#     label_df['SITE_LOCAL'] = pd.NA
#     cond = label_df['ORGAN SITE'] == 'PROSTATE'
#     label_df.loc[cond,'SITE_LOCAL'] = 1
#     label_df.loc[~cond,'SITE_LOCAL'] = 0
    
#     #Rename sample id column
#     label_df.rename(columns = {'TMA-row-col': 'SAMPLE_ID'}, inplace= True)
    
#     #Only select ID that is in label file
#     selected_ids = [x for x in tan_ids if x in list(label_df['SAMPLE_ID'].unique())] #596
    
#     #Exclude IDs has no cancer detected
#     cd_aval_ids = [x for x in os.listdir(info_path) if x != '.DS_Store'] #677
#     cancer_detect_list = []
#     for cur_id in cd_aval_ids:
#         cur_info_df = pd.read_csv(os.path.join(info_path, cur_id, 'ft_model',cur_id + "_TILE_TUMOR_PERC.csv"))
#         cancer_detect_list.append(cur_info_df)
#     all_cd_df = pd.concat(cancer_detect_list) #146888
    
    
#     #Filter for Cancer detected tiles > threshod
#     all_cd_df = all_cd_df.loc[all_cd_df['TUMOR_PIXEL_PERC'] >= args.TUMOR_FRAC_THRES] #19098

#     #No Cancer IDs 
#     cancer_ids = list(set(all_cd_df['SAMPLE_ID']))
#     nocancer_ids = [x for x in cd_aval_ids if x not in cancer_ids] #299
#     print("No Cancer detected (n = ", len(nocancer_ids), ")") 
#     selected_ids = [x for x in selected_ids if x not in  nocancer_ids] 
#     selected_ids.sort()
#     print("Final TCGA IDs (n = ", len(selected_ids), ")") #355
    

#     ############################################################################################################
#     #Combine site and label info and tile info
#     ############################################################################################################     
#     tile_info_list = []
#     for cur_id in selected_ids:
#         cur_info_df = pd.read_csv(os.path.join(info_path, cur_id, 'ft_model',cur_id + "_TILE_TUMOR_PERC.csv"))
#         cur_label_df = label_df.loc[label_df['SAMPLE_ID'] == cur_id]
#         cur_comb_df = cur_info_df.merge(cur_label_df, on = ['SAMPLE_ID'],how = 'left') #add label
#         tile_info_list.append(cur_comb_df)
#     all_tile_info_df = pd.concat(tile_info_list)
#     print(all_tile_info_df.shape) #80630 tiles overlap0
    
#     #Print stats
#     tile_counts = all_tile_info_df['SAMPLE_ID'].value_counts()
#     print("Total OPX IDs in tile path: ", len(set(all_tile_info_df['SAMPLE_ID']))) #355
#     print("Max # tile/per pt:", tile_counts.max()) #279
#     print("Min # tile/per pt:", tile_counts.min()) #54
#     print("Median # tile/per pt:", tile_counts.median()) #236
